ゲーム処理、復帰処理

+++ app/Models/RoomUser.ts
@@ -1,6 +1,7 @@
 import { DateTime } from 'luxon'
 import { BaseModel, column } from '@ioc:Adonis/Lucid/Orm'
 import { TransactionClientContract } from '@ioc:Adonis/Lucid/Database'
+import Room from 'App/Models/Room'
 
 export default class RoomUser extends BaseModel {
   @column({ isPrimary: true })
@@ -24,6 +25,20 @@ export default class RoomUser extends BaseModel {
   }
 
   public static async joinRoom(roomId: number, userId: number, trx: TransactionClientContract) {
+    const room = await Room.query().useTransaction(trx).where('id', roomId).first()
+    if (!room) throw new Error('room not found')
+
+    console.log(room.gameData.players)
+
+    room.gameData.players.push({
+      turnIndex: room.gameData.players.length,
+      playerId: userId,
+      isReady: false,
+      money: 0,
+    })
+    room.useTransaction(trx)
+    await room.save()
+
     const roomUser = new RoomUser()
     roomUser.useTransaction(trx)
     roomUser.roomId = roomId
@@ -33,7 +48,23 @@ export default class RoomUser extends BaseModel {
     return roomUser
   }
 
-  public static async leaveAllRoom(userId: number, trx: TransactionClientContract) {
+  public static async leaveAllRoom(roomId: number, userId: number, trx: TransactionClientContract) {
+    const room = await Room.query().useTransaction(trx).where('id', roomId).first()
+    if (!room) throw new Error('room not found')
+
+    const gameData = room.gameData
+    gameData.players = gameData.players.filter((p) => p.playerId !== userId)
+    room.useTransaction(trx)
+    room.gameData = gameData
+    await room.save()
+
     await RoomUser.query().useTransaction(trx).where('user_id', userId).delete()
+  }
+
+  public static async getByPlayerId(userId: number, trx?: TransactionClientContract) {
+    const query = RoomUser.query()
+    if (trx) query.useTransaction(trx)
+
+    return await query.where('user_id', userId).first()
   }
 }
